Testing

+++ ci/lava/dependencies/enable-wifi.sh
@@ -102,25 +102,30 @@ then
         apt-get install -q -q --yes wget
 
         wget https://www.nxp.com/lgfiles/NMG/MAD/YOCTO/firmware-qca-2.0.3.bin
+        if [ -f firmware-qca-2.0.3.bin ]; then
+            chmod +x firmware-qca-2.0.3.bin
 
-        $mbl_command put firmware-qca-2.0.3.bin /tmp
+            $mbl_command put firmware-qca-2.0.3.bin /tmp
 
-        $mbl_command shell 'ls  /'
-        $mbl_command shell '/tmp/firmware-qca-2.0.3.bin --auto-accept'
-        $mbl_command shell 'cp -v -r /tmp/firmware-qca-2.0.31PJ_QCA9377-3_LEA_2.0/* /'
+            $mbl_command shell 'ls  /'
+            $mbl_command shell '/tmp/firmware-qca-2.0.3.bin --auto-accept'
+            $mbl_command shell 'cp -v -r /tmp/firmware-qca-2.0.31PJ_QCA9377-3_LEA_2.0/* /'
 
-        $mbl_command shell 'ls  /'
-        $mbl_command shell 'ls -R  /lib/modules/'
-        $mbl_command shell 'modprobe qca9733'
-        sleep 60
+            $mbl_command shell 'ls  /'
+            $mbl_command shell 'ls -R  /lib/modules/'
+            $mbl_command shell 'modprobe qca9733'
+            sleep 60
 
-        # Enable WiFi
-        $mbl_command shell 'connmanctl enable wifi'
-        sleep 60
+            # Enable WiFi
+            $mbl_command shell 'connmanctl enable wifi'
+            sleep 60
 
 
-        printf "<LAVA_SIGNAL_TESTCASE TEST_CASE_ID=enable_wifi RESULT=pass>\n"
+            printf "<LAVA_SIGNAL_TESTCASE TEST_CASE_ID=enable_wifi RESULT=pass>\n"
 
+        else
+            printf "<LAVA_SIGNAL_TESTCASE TEST_CASE_ID=enable_wifi RESULT=fail>\n"
+        fi
 
     fi
 else
